ja_erl_utils_binint (ja_erl_utils v1.0.0)

Module for working with binary integers and bits.

Types

binint()

-type binint() :: [bit()].

Represents a binary integer number.

The list of bits starts with the least significant bit and ends with the most significant one.

bit()

-type bit() :: 0 | 1.

Represents one bit.

Functions

and_(BinInt1, BinInt2)

-spec and_(BinInt1 :: binint(), BinInt2 :: binint()) -> BinInt1AndBinInt2 :: binint().

Performs bitwise and operation on two binary integers.

bit_and(Bit1, Bit2)

-spec bit_and(Bit1 :: bit(), Bit2 :: bit()) -> Bit1AndBit2 :: bit().

Performs and operation on two bits.

In addition to Erlang's band operator, it checks if parameters are bits.

bit_invert(Bit)

-spec bit_invert(Bit :: bit()) -> InvertedBit :: bit().

Inverts a single bit.

bit_or(Bit1, Bit2)

-spec bit_or(Bit1 :: bit(), Bit2 :: bit()) -> Bit1OrBit2 :: bit().

Performs or operation on two bits.

In addition to Erlang's bor operator, it checks if parameters are bits.

bit_xor(Bit1, Bit2)

-spec bit_xor(Bit1 :: bit(), Bit2 :: bit()) -> Bit1XorBit2 :: bit().

Performs xor operation on two bits.

In addition to Erlang's bxor operator, it checks if parameters are bits.

drop_insignificant_zeros(BinInt)

-spec drop_insignificant_zeros(BinInt :: binint()) -> Result :: binint().

Drops insignificant 0s (the ones, that are in the most significant positions) of binary integer.

from_integer(Integer)

-spec from_integer(Integer :: integer()) -> BinInt :: binint().

Returns binary representation of integer.

invert(BinInt)

-spec invert(BinInt :: binint()) -> BinIntInverted :: binint().

Inverts bits of a binary integer.

The resulting binary integer might contain 0 as the most significant number. Use drop_insignificant_zeros/1 to remove them.

or_(BinInt1, BinInt2)

-spec or_(BinInt1 :: binint(), BinInt2 :: binint()) -> BinInt1OrBinInt2 :: binint().

Performs bitwise or operation on two binary integers.

to_integer(BinInt)

-spec to_integer(BinInt :: binint()) -> Integer :: integer().

Returns integer from its binary representation.

xor_(BinInt1, BinInt2)

-spec xor_(BinInt1 :: binint(), BinInt2 :: binint()) -> BinInt1XorBinInt2 :: binint().

Performs bitwise xor operation on two binary integers.