ja_erl_utils_int (ja_erl_utils v0.1.0)

Module for working with integer numbers.

Functions

concat(IntStart, IntEnd)

-spec concat(IntStart :: integer(), IntEnd :: integer()) -> IntStartEnd :: integer().

Returns an integer, which results from concatenating second parameter to the end of the first parameter

> ja_erl_utils_int:concat(123, 456).
123456

digit_count(Integer)

-spec digit_count(Integer :: integer()) -> DigitCount :: integer().

Returns the number of digits in non negative integer.

To obtain the result, this function uses math:log10/1, which returns a floating point number. Converting floats to integers might cause errors due to loss of precision. To combat this, math:log10/1 is applied only to sufficiently small numbers (less than 10^12). Larger numbers are divided (using integer division, which is lossless), until they fit this interval.

TODO: Try recursive solution with single digit step, accumulating the result:

fun(Int, Number, Power) when Int < Number -> Power;
fun(Int, Number, Power) _                 -> fun(Int, Number*10, Power+1).

euclidean_div(A, B)

-spec euclidean_div(A :: integer(), B :: integer()) -> ADivB :: integer().

Calculates the quotient of Euclidean division A/B.

Euclidean division

According to Euclid's division lemma, for any two integers A and B, such that B =/= 0, there is a unique pair Q and R such that A = Q*B + R and 0 =< R < |B|. This Q is the quotient of Euclidean division and R is the remainder.

From the lemma it follows that quotient is such integer, that gives the largest product Q*B, which is still not greater than A. If B is positive, then quotient is the largest integer, which is still not greater than the result of floating point division A/B. If B is negative, then quotient is the smallest integer, which is still not smaller than the result of floating point division A/B. In some cases this result differs from the result of Erlang's div operator:

> ja_erl_utils_int:euclidean_div(-7, 3).
-3
2> -7 div 3.
-2
3> ja_erl_utils_int:euclidean_div(-7, -3).
3
4> -7 div -3.
2

euclidean_rem(A, B)

-spec euclidean_rem(A :: integer(), B :: integer()) -> ARemB :: integer().

Calculates the remainder of Euclidean division A/B.

The result is always a non negative integer, which is smaller than absolute value of B. In some cases it differs from the result of Erlang's rem operator:

> ja_erl_utils_int:euclidean_rem(-7, 3).
2
> -7 rem 3.
-1
> ja_erl_utils_int:euclidean_rem(-7, -3).
2
> -7 rem -3.
-1

For the definition of Euclidean division, see euclidean_div/2.

solve_one_equation(Equation)

-spec solve_one_equation(Equation :: {A :: integer(), B :: integer()}) -> integer() | any | undefined.

Solves a single linear equation A*x = B for integer x.

Returns a value of x, if it is the only integer solution of the equation. If there are infinitely many integer solutions (0*x = 0), returns any. If there are no integer solutions (2*x = 3 or 0*x = 3), returns undefined.

> ja_erl_utils_int:solve_one_equation({2, -4}).
-2
> ja_erl_utils_int:solve_one_equation({0, 0}).
any
> ja_erl_utils_int:solve_one_equation({2, 3}).
undefined
> ja_erl_utils_int:solve_one_equation({0, 3}).
undefined

solve_two_equations(Equation1, Equation2)

-spec solve_two_equations(Equation1 :: {A1 :: integer(), B1 :: integer(), C1 :: integer()},
                          Equation2 :: {A2 :: integer(), B2 :: integer(), C2 :: integer()}) ->
                             {X :: integer() | any, Y :: integer() | any | string()} | undefined.

Solves system of two linear equations with two variables for integers.

For equation system

A1*x + B1*y = C1
A2*x + B2*y = C2

returns a tuple of x and y values, if it is the only solution. If there are infinitely many possible integer values for x or y, returns any in that position. If y is dependent on x, returns {any, String}, where String is an arithmetic expression, representing y's dependency on x. Note, that in such case x is not any integer, but any integer, which makes y value (String expression) integer. If the system has no integer solutions, returns undefined.

> ja_erl_utils_int:solve_two_equations({3, 4, 2}, {-2, 1, -5}).
{2,-1}
> ja_erl_utils_int:solve_two_equations({0, 2, 4}, {0, -3, -6}).
{any,2}
> ja_erl_utils_int:solve_two_equations({0, 0, 0}, {1, 0, -2}).
{-2,any}
> ja_erl_utils_int:solve_two_equations({0, 0, 0}, {0, 0, 0}).
{any,any}
> ja_erl_utils_int:solve_two_equations({1, 2, 3}, {-2, -4, -6}).
{any,"(3 - 1*x)/2"}
> ja_erl_utils_int:solve_two_equations({1, -2, 3}, {-1, -3, -5}).
undefined

split(Integer, Split)

-spec split(Integer :: integer(), Split :: integer()) ->
               {IntStart :: integer(), IntEnd :: integer()} | undefined.

Splits integer Integer into two integers, so that the second integer contains Split last digits of the initial integer and the first integer contains the rest digits.

> ja_erl_utils_int:split(12345, 2).
{123,45}

Returns undefined, if Integer does not contain more digits than Split.

ten_pow(Power)

-spec ten_pow(Power :: integer()) -> TenToThePower :: integer().

Returns Power'th power of 10 (10^Power). This function differs from math:pow/2 as it returns an integer.

To obtain the result, this function uses math:pow/2, which returns a floating point number. Converting floats to integers might cause errors due to loss of precision. To combat this, math:pow/2 is applied only to sufficiently small powers (less than 12). Larger powers are calculated by multiplying (using integer multiplication, which is lossless) smaller power, that fit this interval, by 10^12 as many times as needed.

Implementation, using string concatenation and transformation to integer was also considered, but it appeared to be much slower:

erlang:list_to_integer(lists:append(["1"|lists:duplicate(Power, "0")]))

TODO: Try recursive solution with single digit step, accumulating the result:

fun(0,     Number) -> Number;
fun(Power, Number) -> fun(Power-1, Number*10).