ja_erl_utils_list (ja_erl_utils v0.1.0)

Module for working with lists.

Functions

count_elems_sorted(Elem, List)

-spec count_elems_sorted(Elem, List) -> {Count, ListRemaining}
                            when
                                Elem :: ElemType,
                                List :: [ElemType],
                                Count :: integer(),
                                ListRemaining :: [ElemType],
                                ElemType :: term().

Counts number of elements Elem in sorted in ascending order list List.

In addition to the count, returns remaining list with all the Elems and smaller elements removed.

count_elems_start(List)

-spec count_elems_start(List) -> {Count, ListRemaining}
                           when
                               List :: [ElemType],
                               Count :: integer(),
                               ListRemaining :: [ElemType],
                               ElemType :: term().

Counts how many times starting element of list List is repeated at the start of the list.

In addition to the count, returns remaining list with all the copies of starting element removed from the start.

diagonals_b(ListOfLists)

-spec diagonals_b(ListOfLists :: [List]) -> [List] when List :: [term()].

Returns a list of all the backward diagonals (of direction \) of a matrix ListOfLists represented as list of lists.

Diagonals are returned starting from the bottom left corner and ending with the top right corner. Moreover, each returned diagonal is reversed (from the bottom right element to the top left one):

> ja_erl_utils_list:diagonals_b([[1, 2],
                                 [3, 4]]).
[[3],[4,1],[2]]

This function is defined only if each list in ListOfLists is of the same size.

Consider that

This function was created before ja_erl_utils_matrix. It's aim was to ease traversing diagonals of a matrix, which had to be represented as list of lists. The function is probably not that useful now, as it is more effective to walk through matrix diagonals by indexing elements in ja_erl_utils_matrix:matrix/1.

diagonals_f(ListOfLists)

-spec diagonals_f(ListOfLists :: [List]) -> [List] when List :: [term()].

Returns a list of all the forward diagonals (of direction /) of matrix ListOfLists represented as list of lists.

Diagonals are returned in reverse order (starting from the bottom right corner and ending with the top left corner). Moreover, each returned diagonal is reversed (from the bottom left element to the top right one):

> ja_erl_utils_list:diagonals_f([[1, 2],
                                 [3, 4]]).
[[4],[3,2],[1]]

This function is defined only if each list in ListOfLists is of the same size.

See information box in the documentation of diagonals_b/1.

filter_count(Pred, List)

-spec filter_count(Pred, List) -> Count
                      when
                          Pred :: fun((Elem) -> boolean()),
                          Elem :: ElemType,
                          List :: [ElemType],
                          Count :: number(),
                          ElemType :: term().

Filters elements of list List using function (predicate) Pred and counts the remaining elements.

Calling it is equivalent to calling erlang:length(lists:filter(Pred, List)), except that List is traversed only once.

filter_count_foldl(Pred, Acc0, List)

-spec filter_count_foldl(Pred, Acc0, List) -> {Count, Acc1}
                            when
                                Pred :: fun((Elem, AccIn) -> {boolean(), AccOut}),
                                Acc0 :: AccType,
                                Acc1 :: AccType,
                                AccIn :: AccType,
                                AccOut :: AccType,
                                Elem :: ElemType,
                                List :: [ElemType],
                                Count :: number(),
                                ElemType :: term(),
                                AccType :: term().

Combines filter_count/2 and lists:foldl/3.

Calling it is equivalent to calling:

{ListMapped, Acc1} = lists:mapfoldl(Fun, Acc0, List),
{erlang:length(lists:filter(fun(E) -> E end, ListMapped)), Acc1}.

Except that List is traversed only once.

foldl_pairs(Fun, Acc0, List)

-spec foldl_pairs(Fun, Acc0, List) -> Acc1
                     when
                         Fun :: fun((Elem1, Elem2, AccIn) -> AccOut),
                         Acc0 :: AccType,
                         Acc1 :: AccType,
                         AccIn :: AccType,
                         AccOut :: AccType,
                         Elem1 :: ElemType,
                         Elem2 :: ElemType,
                         List :: [ElemType],
                         ElemType :: term(),
                         AccType :: term().

Folds through every pair of different elements of list List.

No element is paired with itself and the order of elements of a pair is not important. If List is [1,2,3,...,n], then the pairs are folded in following order: {1,2}, {1,3}, ..., {1,n}, {2,3}, ..., {2,n},..., {n-1,n}.

intersection(List1, List2)

-spec intersection(List1, List2) -> Intersection
                      when
                          List1 :: [ElemType],
                          List2 :: [ElemType],
                          Intersection :: [ElemType],
                          ElemType :: term().

Finds the intersection of two lists List1 and List2.

That is, it returns a list, containing those elements, which are in both given lists List1 and List2.

is_decreasing(List)

-spec is_decreasing(List :: [number()]) -> boolean().

Checks if list List is in decreasing (non-increasing) order.

Every element of the list must be smaller or equal to the previous one.

is_decreasing(List, Diff)

-spec is_decreasing(List :: [number()], Diff :: number()) -> boolean().

Checks if list List is in decreasing (non-increasing) order with consecutive elements not further apart than Diff.

Every element of the list must be smaller than previous element by not more than Diff or equal to it.

is_increasing(List)

-spec is_increasing(List :: [number()]) -> boolean().

Checks if list List is in increasing (non-decreasing) order.

Every element of the list must be larger or equal to the previous one.

is_increasing(List, Diff)

-spec is_increasing(List :: [number()], Diff :: number()) -> boolean().

Checks if list List is in increasing (non-decreasing) order with consecutive elements not further apart than Diff.

Every element of the list must be larger than previous element by not more than Diff or equal to it.

is_monotonic(List, Fun)

-spec is_monotonic(List, Fun) -> boolean()
                      when List :: [number()], Fun :: fun((number(), number()) -> boolean()).

Checks if list List is monotonic according to Fun.

Checks if for every pair of consecutive elements Prev and Next call to Fun(Prev, Next) returns true.

is_steady(List, Diff)

-spec is_steady(List :: [number()], Diff :: number()) -> boolean().

Checks if consecutive elements of list List are not further apart than Diff.

is_strictly_decreasing(List)

-spec is_strictly_decreasing(List :: [number()]) -> boolean().

Checks if list List is in strictly decreasing (descending) order.

Every element of the list must be strictly smaller than the previous one.

is_strictly_decreasing(List, Diff)

-spec is_strictly_decreasing(List :: [number()], Diff :: number()) -> boolean().

Checks if list List is in strictly decreasing (descending) order with consecutive elements not further apart than Diff.

Every element of the list must be strictly smaller than previous element by not more than Diff.

is_strictly_increasing(List)

-spec is_strictly_increasing(List :: [number()]) -> boolean().

Checks if list List is in strictly increasing (ascending) order.

Every element of the list must be strictly larger than the previous one.

is_strictly_increasing(List, Diff)

-spec is_strictly_increasing(List :: [number()], Diff :: number()) -> boolean().

Checks if list List is in strictly increasing (ascending) order with consecutive elements not further apart than Diff.

Every element of the list must be strictly larger than previous element by not more than Diff.

map_sum(Fun, List)

-spec map_sum(Fun, List) -> Sum
                 when
                     Fun :: fun((Elem) -> number()),
                     Elem :: ElemType,
                     List :: [ElemType],
                     Sum :: number(),
                     ElemType :: term().

Maps every element of list List to a number using function Fun and adds those numbers.

Calling it is equivalent to calling lists:sum(lists:map(Fun, List)), except that List is traversed only once.

map_sum_foldl(Fun, Acc0, List)

-spec map_sum_foldl(Fun, Acc0, List) -> {Sum, Acc1}
                       when
                           Fun :: fun((Elem, AccIn) -> {number(), AccOut}),
                           Acc0 :: AccType,
                           Acc1 :: AccType,
                           AccIn :: AccType,
                           AccOut :: AccType,
                           Sum :: number(),
                           Elem :: ElemType,
                           List :: [ElemType],
                           ElemType :: term(),
                           AccType :: term().

Combines map_sum/2 and lists:foldl/3.

Calling it is equivalent to calling:

{ListMapped, Acc1} = lists:mapfoldl(Fun, Acc0, List),
{lists:sum(ListMapped), Acc1}.

Except that List is traversed only once.

middle(List)

-spec middle(List :: [ElemType]) -> [ElemType] when ElemType :: term().

Returns the middle element(s) of the list List.

If the list contains odd number of elements, the returned list contains a single middle element. If however it contains even number of elements, two elements next to the middle of List is returned.

middle_single(List)

-spec middle_single(List :: [ElemType]) -> ElemType when ElemType :: term().

Returns the single middle element of the list List.

This function is designed for lists, containing odd number of elements. If List contains even number of elements, this function crashes.

transpose(ListOfLists)

-spec transpose(ListOfLists :: [List]) -> [List] when List :: [term()].

Transposes a matrix ListOfLists represented as list of lists.

Transposition is flipping matrix over its diagonal:

> ja_erl_utils_list:transpose([[1, 2],
                               [3, 4]]).
[[1,3],[2,4]]

This function is defined only if each list in ListOfLists is of the same size.