ja_erl_utils_map (ja_erl_utils v1.0.0)

Module for working with maps.

Functions

filter_count(Pred, Map)

-spec filter_count(Pred, Map) -> Count
                      when
                          Pred :: fun((Key, Value) -> boolean()),
                          Map :: #{KeyType => ValueType},
                          Count :: number(),
                          Key :: KeyType,
                          Value :: ValueType,
                          KeyType :: term(),
                          ValueType :: term().

Filters key-value pairs of map Map using function (predicate) Pred and counts the remaining elements.

Calling it is equivalent to calling maps:size(maps:filter(Pred, Map)), except that Map is traversed only once.

filter_count_fold(Pred, Acc0, Map)

-spec filter_count_fold(Pred, Acc0, Map) -> {Count, Acc1}
                           when
                               Pred :: fun((Key, Value, AccIn) -> {boolean(), AccOut}),
                               Map :: #{KeyType => ValueType},
                               Acc0 :: AccType,
                               Acc1 :: AccType,
                               AccIn :: AccType,
                               AccOut :: AccType,
                               Count :: number(),
                               Key :: KeyType,
                               Value :: ValueType,
                               KeyType :: term(),
                               ValueType :: term().

Combines filter_count/2 and maps:fold/3.

> ja_erl_utils_map:filter_count_fold(
    fun(Key, Value, AccIn) -> {Key > Value, AccIn + Key + Value} end,
    0,
    #{1 => 2, 4 => 3, 6 => 0}
  ).
{2,16}

map_sum(Fun, Map)

-spec map_sum(Fun, Map) -> Sum
                 when
                     Fun :: fun((Key, Value) -> number()),
                     Map :: #{KeyType => ValueType},
                     Sum :: number(),
                     Key :: KeyType,
                     Value :: ValueType,
                     KeyType :: term(),
                     ValueType :: term().

Maps every key-value pair of map Map to a number using function Fun and adds those numbers.

Calling it is equivalent to calling lists:sum(maps:values(maps:map(Fun, Map))) except that Map is traversed only once.

map_sum_fold(Fun, Acc0, Map)

-spec map_sum_fold(Fun, Acc0, Map) -> {Sum, Acc1}
                      when
                          Fun :: fun((Key, Value, AccIn) -> {number(), AccOut}),
                          Acc0 :: AccType,
                          Acc1 :: AccType,
                          AccIn :: AccType,
                          AccOut :: AccType,
                          Map :: #{KeyType => ValueType},
                          Sum :: number(),
                          Key :: KeyType,
                          Value :: ValueType,
                          KeyType :: term(),
                          ValueType :: term(),
                          AccType :: term().

Combines map_sum/2 and maps:fold/3.

> ja_erl_utils_map:map_sum_fold(
    fun(Key, Value, AccIn) -> {Key*Value, AccIn + Key + Value} end,
    0,
    #{1 => 2, 4 => 3, 6 => 0}
  ).
{14,16}

max_value(Map)

-spec max_value(Map) -> MaxValue | undefined
                   when
                       Map :: #{KeyType => ValueType},
                       MaxValue :: ValueType,
                       KeyType :: term(),
                       ValueType :: term().

Returns maximum among the values in map Map.

If Map is empty, returns undefined.

min_value(Map)

-spec min_value(Map) -> MaxValue | undefined
                   when
                       Map :: #{KeyType => ValueType},
                       MaxValue :: ValueType,
                       KeyType :: term(),
                       ValueType :: term().

Returns minimum among the values in map Map.

If Map is empty, returns undefined.

reverse(Map)

-spec reverse(Map) -> ReversedMap
                 when
                     Map :: #{KeyType => ValueType},
                     ReversedMap :: #{ValueType => KeyType},
                     KeyType :: term(),
                     ValueType :: term().

Reverses the map Map so that instead of key-value pairs it stores value-key pairs.

This function expects that Map is injective. If however two (or more) keys are associated with the same value in Map, then the value in ReversedMap will be associated to one of those keys, however it is not defined, which one.

> ja_erl_utils_map:reverse(#{a => 1, b => 2}).
#{1 => a,2 => b}