ja_erl_utils_matrix (ja_erl_utils v0.1.0)

Module for working with matrices: tables of data.

For a definition of matrix see matrix/1.

Types

direction()

-type direction() :: up | right | down | left.

Direction to move to the next matrix cell.

index()

-type index() :: {Row :: integer(), Column :: integer()}.

Index of a matrix cell, composing of cell's row Row and column Column.

The index is unique for each cell. Each matrix cell can be effectively retrieved and changed by its index. Indexing of both rows and columns starts at 1.

iteration_opts()

-type iteration_opts() ::
          #{row_order => left | right, column_order => top | bottom, loop_order => rows | columns}.

Options that affect the order of iteration through the matrix.

Possible values are:

row_order (default: left):

left: each row should be iterated from left (the beginning) to right (the end).

right: each row should be iterated from right to left.

column_order (default: top):

top: columns should be iterated from the top (the first one) to the bottom (the last one).

bottom: columns should be iterated from the bottom to the top.

loop_order (default: rows):

rows: iteration through rows: all the columns of single row should be iterated before moving to the next row.

columns: iteration through columns: all the rows of single column should be iterated before moving to the next column.

iterator(Matrix)

-opaque iterator(_Type)

An iterator to iterate through (possibly part of) matrix.

matrix(Type)

-opaque matrix(_Type)

A matrix is a table of data of type Type.

Matrix is composed of cells organized in rows and columns. Each row has the same number of columns and each column has the same number of rows.

Functions

dimensions(Matrix)

-spec dimensions(Matrix) -> Dimensions
                    when
                        Matrix :: matrix(any()),
                        Dimensions :: {Rows, Columns},
                        Rows :: integer(),
                        Columns :: integer().

Returns matrix dimensions: row count Rows and column count Columns.

direction_all()

This function is deprecated. ja_erl_utils_matrix:direction_all/0 is deprecated; Use ?DIRECTIONS_ALL macro of ja_erl_utils_matrix.hrl instead.

-spec direction_all() -> [direction()].

Returns all possible directions.

direction_clockwise(Direction)

-spec direction_clockwise(Direction) -> direction() when Direction :: direction().

Returns direction after a clockwise turn from Direction.

direction_counterclockwise(Direction)

-spec direction_counterclockwise(Direction) -> direction() when Direction :: direction().

Returns direction after a counterclockwise turn from Direction.

direction_reverse(Direction)

-spec direction_reverse(Direction) -> direction() when Direction :: direction().

Returns opposite direction of Direction.

filter_count(Pred, Matrix)

-spec filter_count(Pred, Matrix) -> Count
                      when
                          Pred :: fun((Index, Value) -> boolean()),
                          Matrix :: matrix(Type),
                          Index :: index(),
                          Value :: Type,
                          Count :: number(),
                          Type :: any().

Filters cells of matrix Matrix using function (predicate) Pred and counts the remaining ones.

If Matrix represents matrix

1 2 3
4 5 6

then

> ja_erl_utils_matrix:filter_count(
    fun({Row, Column}, Value) -> Row =:= 2 orelse Column =:= 2 orelse Value rem 3 =:= 0 end,
    Matrix
  ).
5

filter_count_foldf(Pred, Acc0, Matrix)

-spec filter_count_foldf(Pred, Acc0, Matrix) -> {Count, Acc1}
                            when
                                Pred :: fun((Index, Value, AccIn) -> {boolean(), AccOut}),
                                Acc0 :: AccType,
                                Acc1 :: AccType,
                                AccIn :: AccType,
                                AccOut :: AccType,
                                Matrix :: matrix(Type),
                                Index :: index(),
                                Value :: Type,
                                Count :: number(),
                                Type :: any(),
                                AccType :: any().

Combines filter_count/2 and foldf/3.

If Matrix represents matrix

1 2 3
4 5 6

then

> ja_erl_utils_matrix:filter_count_foldf(
    fun({Row, Column}, Value, Acc) -> {Value > 4, Acc + Row + Column + Value} end,
    0,
    Matrix
  ).
{2, 42}

fold(Fun, Acc0, Matrix, Opts)

-spec fold(Fun, Acc0, Matrix, Opts) -> Acc1
              when
                  Fun :: fun((Index, Value, AccIn) -> AccOut),
                  Acc0 :: AccType,
                  Acc1 :: AccType,
                  AccIn :: AccType,
                  AccOut :: AccType,
                  Matrix :: matrix(Type),
                  Value :: Type,
                  Index :: index(),
                  Opts :: iteration_opts() | #{loop_end_fun => fun((RowOrColumn, AccIn) -> AccOut)},
                  RowOrColumn :: integer(),
                  AccType :: any(),
                  Type :: any().

Folds through cells of matrix Matrix using function Fun and initial accumulator Acc0. Order of fold is defined in Opts.

In addition to options defined in iteration_opts/0, the following option in Opts is possible:

loop_end_fun: how accumulator must be altered after traversing full row (if loop_order is rows) or full column (if loop_order is columns). The function is passed the index of the row (if loop_order is rows) or column (if loop_order is columns), which has just been traversed. By default the same unaltered accumulator is returned.

foldb(Fun, Acc0, Matrix)

-spec foldb(Fun, Acc0, Matrix) -> Acc1
               when
                   Fun :: fun((Index, Value, AccIn) -> AccOut),
                   Acc0 :: AccType,
                   Acc1 :: AccType,
                   AccIn :: AccType,
                   AccOut :: AccType,
                   Matrix :: matrix(Type),
                   Index :: index(),
                   Value :: Type,
                   AccType :: any(),
                   Type :: any().

Folds back through cells of matrix Matrix using function Fun and initial accumulator Acc0.

The fold is started from the last row and last column. Each column of the row is handled consecutively before going to previous row.

foldf(Fun, Acc0, Matrix)

-spec foldf(Fun, Acc0, Matrix) -> Acc1
               when
                   Fun :: fun((Index, Value, AccIn) -> AccOut),
                   Acc0 :: AccType,
                   Acc1 :: AccType,
                   AccIn :: AccType,
                   AccOut :: AccType,
                   Matrix :: matrix(Type),
                   Index :: index(),
                   Value :: Type,
                   AccType :: any(),
                   Type :: any().

Folds forward through cells of matrix Matrix using function Fun and initial accumulator Acc0.

The fold is started from first row and first column. Each column of the row is handled consecutively before going to next row.

get(Index, Matrix)

-spec get(Index, Matrix) -> Value
             when Index :: index(), Matrix :: matrix(Type), Value :: Type, Type :: any().

Returns value of the cell at index Index of matrix Matrix.

Throws badarg error, if Index is not in Matrix.

get_char_matrix(Lines)

-spec get_char_matrix(Lines) -> Matrix when Lines :: [Line :: string()], Matrix :: matrix(char()).

Converts list of lines Lines to a matrix, with each character as a separate cell.

For each character, its line position in Lines is the row part of cell's index. Character's position in its line is the column part of cell's index. Each line in Lines must contain the same number of elements.

index_of(Value, Matrix)

-spec index_of(Value, Matrix) -> index() | undefined
                  when Value :: Type, Matrix :: matrix(Type), Type :: any().

Returns the index of cell with value Value in matrix Matrix.

If Value is not found, undefined is returned. If there are several occurrences of Value in Matrix, one of the indexes is returned, but it is not specified, which one.

is_valid_index(Index, Matrix)

-spec is_valid_index(Index, Matrix) -> boolean() when Index :: index(), Matrix :: matrix(any()).

Returns true if index Index is in matrix Matrix. Otherwise returns false.

Index is in Matrix, if both column and row part of Index are positive, row part is not larger that row count of Matrix and column part is not larger than column count of Matrix.

iterator(Matrix)

-spec iterator(Matrix) -> Iterator
                  when Matrix :: matrix(Type), Iterator :: iterator(Type), Type :: any().

Returns matrix Matrix iterator with default iteration options.

Equivalent to calling iterator(Matrix, #{}).

iterator(Matrix, Opts)

-spec iterator(Matrix, Opts) -> Iterator
                  when
                      Matrix :: matrix(Type),
                      Opts :: iteration_opts(),
                      Iterator :: iterator(Type),
                      Type :: any().

Returns matrix Matrix iterator with iteration options Opts.

Iterator can be used to iterate through matrix using next/1 function.

map(Fun, Matrix1)

-spec map(Fun, Matrix1) -> Matrix2
             when
                 Fun :: fun((Index, Value1) -> Value2),
                 Matrix1 :: matrix(Type),
                 Matrix2 :: matrix(Type),
                 Index :: index(),
                 Value1 :: Type,
                 Value2 :: Type,
                 Type :: any().

Maps every cell of matrix Matrix using function Fun and returns new matrix with mapped values.

map_sum(Fun, Matrix)

-spec map_sum(Fun, Matrix) -> Sum
                 when
                     Fun :: fun((Index, Value) -> number()),
                     Matrix :: matrix(Type),
                     Index :: index(),
                     Value :: Type,
                     Sum :: number(),
                     Type :: any().

Maps every cell of matrix Matrix to a number using function Fun and adds those numbers.

Calling it is equivalent to calling lists:sum(ja_erl_utils_matrix:values(ja_erl_utils_matrix:map(Fun, Matrix))), except that Matrix is traversed only once.

map_sum_foldf(Fun, Acc0, Matrix)

-spec map_sum_foldf(Fun, Acc0, Matrix) -> {Sum, Acc1}
                       when
                           Fun :: fun((Index, Value, AccIn) -> {number(), AccOut}),
                           Acc0 :: AccType,
                           Acc1 :: AccType,
                           AccIn :: AccType,
                           AccOut :: AccType,
                           Matrix :: matrix(Type),
                           Sum :: number(),
                           Index :: index(),
                           Value :: Type,
                           Type :: term(),
                           AccType :: term().

Combines map_sum/2 and foldf/3.

If Matrix represents matrix

1 2
3 4

then

> ja_erl_utils_matrix:map_sum_foldf(
    fun({Row, Column}, Value, AccIn) -> {Row*Column*Value, AccIn + Row + Column + Value} end,
    0,
    Matrix
  ).
{27,22}

new(Rows, Columns)

-spec new(Rows, Columns) -> matrix(undefined) when Rows :: integer(), Columns :: integer().

Returns a new matrix with Rows rows and Columns columns and with all cells filled by undefined.

Calling it is equivalent to calling ja_erl_utils_matrix:new(Rows, Columns, undefined).

new(Rows, Columns, Default)

-spec new(Rows, Columns, Default) -> matrix(Type)
             when Rows :: integer(), Columns :: integer(), Default :: Type, Type :: any().

Returns a new matrix with Rows rows and Columns columns and with all the cells filled by Default.

next(Iterator)

-spec next(Iterator) -> {Index, Value, NextIterator} | none
              when
                  Iterator :: iterator(Type),
                  NextIterator :: iterator(Type),
                  Index :: index(),
                  Value :: Type,
                  Type :: any().

Returns next index and value according to iterator Iterator and a new iterator for remaining data.

If there are no data according to iterator, returns none.

> Matrix = ja_erl_utils_matrix:get_char_matrix(["ab","cd"]),
  It = ja_erl_utils_matrix:iterator(Matrix), ok.
ok
> {I1, V1, It1} = ja_erl_utils_matrix:next(It), {I1, V1}.
{{1,1},97}
> {I2, V2, It2} = ja_erl_utils_matrix:next(It1), {I2, V2}.
{{1,2},98}
> {I3, V3, It3} = ja_erl_utils_matrix:next(It2), {I3, V3}.
{{2,1},99}
> {I4, V4, It4} = ja_erl_utils_matrix:next(It3), {I4, V4}.
{{2,2},100}
> ja_erl_utils_matrix:next(It4).
none

next_index(Index, Direction, Matrix)

-spec next_index(Index, Direction, Matrix) -> index() | undefined
                    when Index :: index(), Direction :: direction(), Matrix :: matrix(any).

Returns index of the next cell of matrix Matrix, starting with index Index in direction Direction.

If there is no cell in that direction (it is out of matrix bounds), undefined is returned.

print_char_matrix(Matrix)

-spec print_char_matrix(Matrix) -> ok when Matrix :: matrix(char()).

Outputs matrix Matrix of characters to terminal.

set(Index, Matrix1, Value)

-spec set(Index, Matrix1, Value) -> Matrix2
             when
                 Index :: index(),
                 Matrix1 :: matrix(Type),
                 Matrix2 :: matrix(Type),
                 Value :: Type,
                 Type :: any().

Sets value of the cell at index Index of matrix Matrix to Value.

Throws badarg error, if Index is not in Matrix.

values(Matrix)

-spec values(Matrix) -> Values when Matrix :: matrix(Type), Values :: [Type], Type :: any().

Returns list of all the cell values of matrix Matrix.