ja_erl_utils_term (ja_erl_utils v1.0.0)

Module for functions, that can be used on any Erlang term.

Some functions of this module were part of other projects. For example, format/1 and format/2 functions are based on erlang_tools_display_term_erl:to_binary/1,2 functions of ErlangTools project.

Types

position()

-type position() :: [position_level()].

Position (address) inside Erlang's term.

Position is a list of indexes for levels of term's depth. The first element of the list corresponds to the outermost level of the term, the second element - to the second level, etc. If the term doesn't have required depth in required places, then the position doesn't address anything in it.

Empty position [] doesn't address any part of the term.

Here are some examples of position:

[2] addresses the second element of any list or tuple and map entry with key 2: atom this in [a, this, b, c] or {a, this, b} and 2 => this in #{1 => a, 2 => this}.

[key] addresses map entry with key key: key => this in #{a => b, key => this}.

[3, 2, key, 1]: if term is a list or tuple, whose third element is a list or tuple, whose second element is a map, which has entry with 'key' as a key and a list or tuple as a value, then this position addresses the first element of the last list or tuple: atom this in term [a, [b,c], {d, #{e => f, key => [this, g, h, i]}, j}].

[2, [2, 4]]: if term is a list or tuple, whose second element is a list or tuple, then this position addresses the second and the fourth elements of the latter list or tuple: atoms this in term [a, {b, this, c, this, d}].

[2, '*']: if term is a list or tuple, whose second element is a list or tuple, then this position addresses all the elements of the latter list or tuple: atoms this in term {a, [this, this, this]}.

[2, {2, 4}]: if term is a list or tuple, whose second element is a list or tuple, then this position addresses elements starting with second and ending with fourth of the latter list or tuple: atoms this in term [a, {b, this, this, this, c}].

position_level()

-type position_level() :: position_level_index() | [position_level_index()].

Address in single level of a position.

It can consist of one or several indexes. If it consists of several indexes, it addresses all the elements addressed by at least one of the indexes. For example, [1, 3] addresses first and third elements of list or tuple at that level. [2, {5, 8}] addresses second element and all the elements from fifth to eighth.

position_level_id()

-type position_level_id() :: integer() | term().

Id of element in list, tuple or map.

The type of id depends on the structure:

In a list or tuple, elements are identified by integer, which denotes element's position in the structure. 1 is used to identify the first element.

In a map, elements (map entries) are identified by key. It can also be integer, if such key is part of a map. However, it can be any term as map keys are not limited to integers in Erlang. Note that id identifies the whole entry: both key and value (not only the value). Currently there is no way to identify key only or value only.

position_level_index()

-type position_level_index() ::
          position_level_id() | '*' | {From :: position_level_id(), To :: position_level_id()}.

Atomic address of element(s) in single level of a position.

The values can be:

Id of single element,

'*' meaning all the elements of list, tuple or map at that level,

{From, To} meaning all the elements starting from id From and ending with id End (including End).

Functions

format(Term)

-spec format(Term :: term()) -> binary().

Converts term to pretty formatted binary.

Equivalent to calling ja_erl_utils_term:format(Term, #{}).

format(Term, Options)

-spec format(Term, Options) -> binary()
                when
                    Term :: term(),
                    Options ::
                        #{single_indent_length => integer(),
                          no_expansion => [position()] | true,
                          expand_map_keys => boolean()}.

Converts term Term to pretty formatted binary parametrized by Options.

Pretty formatting is designed with aim to be readable for humans. All the list entries, tuple elements and key-value pairs of map are displayed on a separate line with larger indentation than term, which contains the list, tuple or map. The exception is lists, tuples and maps, which contain exactly one element (the element is not displayed on a separate line) or none elements at all (displayed in a single line).

Possible keys of Options are:

single_indent_length (default: 2) - number of spaces, used in single indent level.

no_expansion (default: []) - list of positions of terms in Term, that should not be expanded to separate lines. For every position in this list, term addressed by it will be displayed on a single line, ignoring the amount of nested lists, tuples and maps in the addressed term.

If value true is provided instead of a list, it means that Term itself should not be expanded. Compare:

  > Term = {[first, second], #{1 => 2, 3 => 4}}.
  {[first,second],#{1 => 2,3 => 4}}
  > ja_erl_utils_terminal:print("~s", [ja_erl_utils_term:format(Term, #{no_expansion => [['*']]})]).
  {
    [first, second],
    #{1 => 2, 3 => 4}
  }
  ok
  > ja_erl_utils_terminal:print("~s", [ja_erl_utils_term:format(Term, #{no_expansion => true})]).
  {[first, second], #{1 => 2, 3 => 4}}
  ok

Note that no_expansion may be a list of positions and never a single position not wrapped in a list. As position itself is a list, this results in quite strange list in a list value [[*]].

expand_map_keys (default: false) - wether lists, tuples and maps should be expanded, if they are part of a key in a map. Usually keys of a map are quite small and although in some cases they might be expandable, it is usually preferable to display them in one line. E.g., if keys of a map are coordinates in two dimensional space ({X, Y} or #{x => X, y => Y}) it would be more readable if they are displayed in one line and values of such map might be expanded to separate lines. Compare:

  > Term = #{{1, 2} => {first, second, third}, {3, 4} => [another, entry]}.
  #{{1,2} => {first,second,third},{3,4} => [another,entry]}
  > ja_erl_utils_terminal:print("~s", [ja_erl_utils_term:format(Term, #{expand_map_keys => false})]).
  #{
    {1, 2} => {
      first,
      second,
      third
    },
    {3, 4} => [
      another,
      entry
    ]
  }
  ok
  > ja_erl_utils_terminal:print("~s", [ja_erl_utils_term:format(Term, #{expand_map_keys => true})]).
  #{
    {
      1,
      2
    } => {
      first,
      second,
      third
    },
    {
      3,
      4
    } => [
      another,
      entry
    ]
  }
  ok

memory_usage(Term)

-spec memory_usage(Term :: term()) ->
                      integer() | {at_least, integer()} | {between, integer(), integer()}.

Estimates how much memory (in bytes) is used to store Erlang's term Term.

This function is based on an article about memory usage in Erlang's documentation, which describes erts-8.0 system in OTP 19.0. As the article itself is not very precise, this function may return:

an integer, which means exact number of bytes as per article;

{at_least, Value} - a term consumes at least Value number of bytes. This is needed, because memory usage of large integers is not specified accurately;

{between, From, To} - a term consumes at least From and at most To number of bytes. This is needed, because memory usage of binaries, references, functions, etc... is specified as "From-To words".

The returned values are always divisible by word size of the system.

This function does not take into account:

memory sharing of binaries (or other types) - each sub-term is counted independently and then the values are added;

memory, consumed by atom, node, process, port or function tables;

memory, consumed by emulator internal data structures used for references;

The article does not fully cover bitstrings, which are not binaries. This function assumes that bitstring consumes the least possible number of full words. E.g., if word takes 8 bytes, then 1-64 bit length bitstrings consume 1 word (8 bytes), 65-128 bit length bitstrings - 2 words (16 bytes), etc.

The article doesn't specify exactly, what it means by function environment. This function evaluates the term (list) returned by calling erlang:fun_info(Term, env) using the same rules as for any other term.

Finally, calling erts_debug:flat_size(Term) or erts_debug:size(Term) will probably get more accurate result. However, these functions are hidden, not documented and considered to be for internal use only.